Example: "Fixed" the deadlocks in concurrent_doubly_linked_list by using std::recursive_mutex.
Ideally, the code should be refactored to avoid locking recursively, but the purpose of the example is to add synchronization with minimal refactoring.

In addition, avoiding recursive locking is primarily a performance optimization.
This means we really shouldn't worry about it before we have a benchmark showing that it matters.